Rename utils.py to retry.py
